+++ pyjoulescope_driver/entry_points/statistics.py
@@ -1,4 +1,4 @@
-# Copyright 2022 Jetperch LLC
+# Copyright 2022-2024 Jetperch LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,14 +13,12 @@
 # limitations under the License.
 
 from pyjoulescope_driver import Driver, time64
+import sys
 import time
 
 
 def parser_config(p):
     """Display Joulescope measurement statistics."""
-    p.add_argument('--verbose', '-v',
-                   action='store_true',
-                   help='Display verbose information.')
     p.add_argument('--js110_host',
                    action='store_true',
                    help='Use JS110 host-side statistics (on-instrument by default).')
@@ -29,7 +27,6 @@ def parser_config(p):
                    type=float,
                    help='The desired statistics frequency.')
     p.add_argument('--duration',
-                   default=1.0,
                    type=time64.duration_to_seconds,
                    help='The capture duration in float seconds. '
                         + 'Add a suffix for other units: s=seconds, m=minutes, h=hours, d=days')
@@ -102,4 +99,4 @@ def on_cmd(args):
                     break
                 time.sleep(0.025)
         except KeyboardInterrupt:
-            pass
+            sys.stdout.flush()